// src/types/vault-item-list.ts
export interface GetValutItemListTypes {
  address: string;
  city: string;
  collectible_id: string;
  country: string;
  created_on: string;
  dbCollectible: {
    item_current_location: string;
  };
  email: string;
  network_id: string;
  order_status: string;
  phone_number: string;
  receiver_wallet_address: string;
  state: string;
  zip_code: string;
  _id: string;
  message: string;
  status: boolean;
}